Split onboarding in multiple files
