+++ Android/EasyBudget/app/src/main/java/com/benoitletondor/easybudgetapp/view/onboarding/OnboardingView.kt
@@ -4,69 +4,39 @@ import android.Manifest
 import android.os.Build
 import android.os.Parcelable
 import androidx.activity.compose.BackHandler
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.calculateEndPadding
 import androidx.compose.foundation.layout.calculateStartPadding
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.imePadding
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.pager.HorizontalPager
-import androidx.compose.foundation.pager.PagerState
 import androidx.compose.foundation.pager.rememberPagerState
-import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.derivedStateOf
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.runtime.snapshotFlow
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.focus.FocusRequester
-import androidx.compose.ui.focus.focusRequester
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalLayoutDirection
-import androidx.compose.ui.platform.LocalSoftwareKeyboardController
-import androidx.compose.ui.res.colorResource
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.TextRange
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.input.TextFieldValue
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.benoitletondor.easybudgetapp.R
-import com.benoitletondor.easybudgetapp.helper.CurrencyHelper
 import com.benoitletondor.easybudgetapp.helper.launchCollect
-import com.benoitletondor.easybudgetapp.view.selectcurrency.SelectCurrencyView
+import com.benoitletondor.easybudgetapp.view.onboarding.subviews.OnboardingPageAccountAmount
+import com.benoitletondor.easybudgetapp.view.onboarding.subviews.OnboardingPageCurrency
+import com.benoitletondor.easybudgetapp.view.onboarding.subviews.OnboardingPageEnd
+import com.benoitletondor.easybudgetapp.view.onboarding.subviews.OnboardingPagePushNotifications
+import com.benoitletondor.easybudgetapp.view.onboarding.subviews.OnboardingPageWelcome
 import com.google.accompanist.permissions.ExperimentalPermissionsApi
 import com.google.accompanist.permissions.PermissionState
 import com.google.accompanist.permissions.PermissionStatus
@@ -74,7 +44,6 @@ import com.google.accompanist.permissions.isGranted
 import com.google.accompanist.permissions.rememberPermissionState
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.first
 import kotlinx.parcelize.Parcelize
 import kotlinx.serialization.Serializable
 import java.lang.IllegalStateException
@@ -105,7 +74,7 @@ fun OnboardingView(
     )
 }
 
-private fun pageIndexToOnboardingPage(index: Int): OnboardingViewModel.OnboardingPage {
+fun pageIndexToOnboardingPage(index: Int): OnboardingViewModel.OnboardingPage {
     val isAndroid33OrMore = Build.VERSION.SDK_INT >= 33
 
     return when(index) {
@@ -257,454 +226,3 @@ private fun OnboardingView(
         }
     )
 }
-
-@Composable
-private fun OnboardingPageWelcome(
-    contentPadding: PaddingValues,
-    onNextPressed: () -> Unit,
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(color = colorResource(R.color.easy_budget_green))
-            .padding(contentPadding),
-        horizontalAlignment = Alignment.CenterHorizontally,
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .weight(1f)
-                .verticalScroll(rememberScrollState())
-                .padding(bottom = 20.dp),
-            horizontalAlignment = Alignment.CenterHorizontally,
-        ) {
-            Image(
-                painter = painterResource(R.drawable.ic_launcher),
-                contentDescription = null,
-            )
-
-            Spacer(modifier = Modifier.height(10.dp))
-
-            Text(
-                modifier = Modifier.fillMaxWidth(),
-                text = stringResource(R.string.onboarding_screen_1_title),
-                color = Color.White,
-                fontSize = 40.sp,
-                textAlign = TextAlign.Center,
-                lineHeight = 46.sp,
-            )
-
-            Spacer(modifier = Modifier.height(60.dp))
-
-            Text(
-                modifier = Modifier.fillMaxWidth(),
-                text = stringResource(R.string.onboarding_screen_1_message),
-                color = Color.White,
-                fontSize = 16.sp,
-                textAlign = TextAlign.Center,
-            )
-
-            Spacer(modifier = Modifier.height(40.dp))
-
-            Text(
-                modifier = Modifier.fillMaxWidth(),
-                text = stringResource(R.string.onboarding_screen_1_message2),
-                color = Color.White,
-                fontSize = 30.sp,
-                textAlign = TextAlign.Center,
-            )
-        }
-
-        Button(
-            colors = ButtonDefaults.buttonColors(
-                containerColor = MaterialTheme.colorScheme.secondary,
-                contentColor = MaterialTheme.colorScheme.onPrimary,
-            ),
-            onClick = onNextPressed,
-        ) {
-            Text(
-                text = stringResource(R.string.onboarding_screen_1_cta),
-                fontSize = 20.sp,
-            )
-        }
-    }
-}
-
-@Composable
-private fun OnboardingPageCurrency(
-    contentPadding: PaddingValues,
-    userCurrencyFlow: StateFlow<Currency>,
-    onNextPressed: () -> Unit,
-) {
-    val selectedCurrency by userCurrencyFlow.collectAsState()
-
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(color = colorResource(R.color.secondary))
-            .padding(contentPadding),
-        horizontalAlignment = Alignment.CenterHorizontally,
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .weight(1f)
-                .padding(bottom = 20.dp),
-            horizontalAlignment = Alignment.CenterHorizontally,
-        ) {
-            Text(
-                modifier = Modifier.fillMaxWidth(),
-                text = stringResource(R.string.onboarding_screen_2_title),
-                color = Color.White,
-                fontSize = 30.sp,
-                textAlign = TextAlign.Center,
-                lineHeight = 36.sp,
-            )
-
-            Spacer(modifier = Modifier.height(30.dp))
-
-            Text(
-                modifier = Modifier.fillMaxWidth(),
-                text = stringResource(R.string.onboarding_screen_2_message),
-                color = Color.White,
-                fontSize = 20.sp,
-                textAlign = TextAlign.Center,
-            )
-
-            SelectCurrencyView(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .weight(weight = 1f, fill = false)
-                    .padding(horizontal = 20.dp, vertical = 30.dp)
-                    .background(color = Color.White),
-            )
-        }
-
-        Button(
-            onClick = onNextPressed,
-        ) {
-            Text(
-                text = stringResource(R.string.onboarding_screen_2_cta, selectedCurrency.symbol),
-                fontSize = 20.sp,
-            )
-        }
-    }
-}
-
-@Composable
-private fun OnboardingPageAccountAmount(
-    contentPadding: PaddingValues,
-    pagerState: PagerState,
-    userCurrencyFlow: StateFlow<Currency>,
-    userMoneyAmountFlow: StateFlow<Double>,
-    onNextPressed: () -> Unit,
-    onAmountChange: (String) -> Unit,
-) {
-    val currency by userCurrencyFlow.collectAsState()
-    val currentAmount by userMoneyAmountFlow.collectAsState()
-
-    var currentTextFieldValue by remember { mutableStateOf(TextFieldValue(
-        text = "",
-        selection = TextRange(index = 0),
-    )) }
-
-    LaunchedEffect("initAmount") {
-        val amountFromDB = userMoneyAmountFlow.value
-        val formattedAmount = formatAmountValue(amountFromDB)
-        if (amountFromDB != 0.0 && formattedAmount != currentTextFieldValue.text) {
-            currentTextFieldValue = TextFieldValue(
-                text = formattedAmount,
-                selection = TextRange(index = formattedAmount.length),
-            )
-        }
-    }
-
-    val focusRequester = remember { FocusRequester() }
-    val keyboardController = LocalSoftwareKeyboardController.current
-
-    val shouldRequestFocus by remember {
-        derivedStateOf {
-            pagerState.currentPageOffsetFraction == 0f && pageIndexToOnboardingPage(pagerState.currentPage) == OnboardingViewModel.OnboardingPage.INITIAL_AMOUNT
-        }
-    }
-
-    LaunchedEffect(shouldRequestFocus) {
-        if (shouldRequestFocus) {
-            focusRequester.requestFocus()
-            keyboardController?.show()
-        } else {
-            focusRequester.freeFocus()
-            keyboardController?.hide()
-        }
-    }
-
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(color = colorResource(R.color.secondary))
-            .padding(contentPadding),
-        horizontalAlignment = Alignment.CenterHorizontally,
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .weight(1f)
-                .verticalScroll(rememberScrollState())
-                .padding(bottom = 20.dp),
-            horizontalAlignment = Alignment.CenterHorizontally,
-        ) {
-            Text(
-                modifier = Modifier.fillMaxWidth(),
-                text = stringResource(R.string.onboarding_screen_3_title),
-                color = Color.White,
-                fontSize = 30.sp,
-                textAlign = TextAlign.Center,
-                lineHeight = 36.sp,
-            )
-
-            Spacer(modifier = Modifier.height(30.dp))
-
-            Text(
-                modifier = Modifier.fillMaxWidth(),
-                text = stringResource(R.string.onboarding_screen_3_message),
-                color = Color.White,
-                fontSize = 20.sp,
-                textAlign = TextAlign.Center,
-            )
-
-            Spacer(modifier = Modifier.height(20.dp))
-
-            Row(
-                modifier = Modifier.fillMaxWidth(0.7f),
-                horizontalArrangement = Arrangement.Center,
-                verticalAlignment = Alignment.CenterVertically,
-            ){
-                TextField(
-                    modifier = Modifier
-                        .weight(1f)
-                        .focusRequester(focusRequester),
-                    value = currentTextFieldValue,
-                    onValueChange = { newValue ->
-                        if (newValue.text.all { "-0123456789.,".contains(it) }) {
-                            val newText = if (newValue.text.count() > 1 && newValue.text.endsWith("-")) {
-                                if (newValue.text.startsWith("-")) {
-                                    newValue.text.substring(1, newValue.text.length - 1)
-                                } else {
-                                    "-${newValue.text.substring(0, newValue.text.length - 1)}"
-                                }
-                            } else {
-                                newValue.text
-                            }
-
-                            currentTextFieldValue = TextFieldValue(
-                                text = newText,
-                                selection = newValue.selection,
-                            )
-                            onAmountChange(newText)
-                        }
-                    },
-                    textStyle = TextStyle(
-                        fontSize = 20.sp,
-                    ),
-                    singleLine = true,
-                    keyboardOptions = KeyboardOptions(
-                        keyboardType = KeyboardType.Number,
-                        autoCorrectEnabled = false,
-                    ),
-                )
-
-                Spacer(modifier = Modifier.width(8.dp))
-
-                Text(
-                    text = currency.symbol,
-                    color = Color.White,
-                    fontSize = 30.sp,
-                )
-            }
-
-        }
-
-        Button(
-            onClick = onNextPressed,
-        ) {
-            Text(
-                text = stringResource(R.string.onboarding_screen_3_cta, CurrencyHelper.getFormattedCurrencyString(currency, currentAmount)),
-                fontSize = 20.sp,
-            )
-        }
-    }
-}
-
-private fun formatAmountValue(amount: Double): String = if (amount == 0.0) "0" else amount.toString()
-
-@Composable
-private fun OnboardingPagePushNotifications(
-    contentPadding: PaddingValues,
-    onAcceptNotificationsPressed: () -> Unit,
-    onDenyNotificationsPressed: () -> Unit,
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(color = colorResource(R.color.secondary))
-            .padding(contentPadding),
-        horizontalAlignment = Alignment.CenterHorizontally,
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .weight(1f)
-                .verticalScroll(rememberScrollState())
-                .padding(bottom = 20.dp),
-            horizontalAlignment = Alignment.CenterHorizontally,
-        ) {
-            Image(
-                painter = painterResource(R.drawable.ic_baseline_notification_important_24),
-                contentDescription = null,
-            )
-
-            Spacer(modifier = Modifier.height(10.dp))
-
-            Text(
-                modifier = Modifier.fillMaxWidth(),
-                text = stringResource(R.string.onboarding_screen_push_permission_title),
-                color = Color.White,
-                fontSize = 30.sp,
-                textAlign = TextAlign.Center,
-                lineHeight = 36.sp,
-            )
-
-            Spacer(modifier = Modifier.height(20.dp))
-
-            Text(
-                modifier = Modifier.fillMaxWidth(),
-                text = stringResource(R.string.onboarding_screen_push_permission_message),
-                color = Color.White,
-                fontSize = 18.sp,
-                textAlign = TextAlign.Center,
-            )
-        }
-
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-        ) {
-            Button(
-                modifier = Modifier.weight(0.5f),
-                onClick = onDenyNotificationsPressed,
-                colors = ButtonDefaults.buttonColors(
-                    containerColor = colorResource(R.color.accent_ripple),
-                    contentColor = colorResource(R.color.easy_budget_green_dark),
-                ),
-            ) {
-                Text(
-                    text = stringResource(R.string.onboarding_screen_push_permission_not_now_cta),
-                    fontSize = 20.sp,
-                )
-            }
-
-            Spacer(modifier = Modifier.width(10.dp))
-
-            Button(
-                modifier = Modifier.weight(0.5f),
-                onClick = onAcceptNotificationsPressed,
-            ) {
-                Text(
-                    text = stringResource(R.string.onboarding_screen_push_permission_accept_cta),
-                    fontSize = 20.sp,
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun OnboardingPageEnd(
-    contentPadding: PaddingValues,
-    onNextPressed: () -> Unit,
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(color = colorResource(R.color.easy_budget_green))
-            .padding(contentPadding),
-        horizontalAlignment = Alignment.CenterHorizontally,
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .weight(1f)
-                .verticalScroll(rememberScrollState())
-                .padding(bottom = 20.dp),
-            horizontalAlignment = Alignment.CenterHorizontally,
-        ) {
-            Image(
-                painter = painterResource(R.drawable.ic_done_white_48dp),
-                contentDescription = null,
-            )
-
-            Spacer(modifier = Modifier.height(10.dp))
-
-            Text(
-                modifier = Modifier.fillMaxWidth(),
-                text = stringResource(R.string.onboarding_screen_4_title),
-                color = Color.White,
-                fontSize = 30.sp,
-                textAlign = TextAlign.Center,
-                lineHeight = 36.sp,
-            )
-
-            Spacer(modifier = Modifier.height(5.dp))
-
-            Text(
-                modifier = Modifier.fillMaxWidth(),
-                text = stringResource(R.string.onboarding_screen_4_message),
-                color = Color.White,
-                fontSize = 18.sp,
-                textAlign = TextAlign.Center,
-            )
-
-            Spacer(modifier = Modifier.height(50.dp))
-
-            Text(
-                modifier = Modifier.fillMaxWidth(),
-                text = stringResource(R.string.onboarding_screen_4_message2),
-                color = Color.White,
-                fontSize = 20.sp,
-                textAlign = TextAlign.Center,
-            )
-
-            Spacer(modifier = Modifier.height(10.dp))
-
-            Text(
-                modifier = Modifier.fillMaxWidth(),
-                text = stringResource(R.string.onboarding_screen_4_message3),
-                color = Color.White,
-                fontSize = 18.sp,
-            )
-
-            Spacer(modifier = Modifier.height(50.dp))
-
-            Text(
-                modifier = Modifier.fillMaxWidth(),
-                text = stringResource(R.string.onboarding_screen_4_message4),
-                color = Color.White,
-                fontSize = 20.sp,
-                textAlign = TextAlign.Center,
-            )
-        }
-
-        Button(
-            colors = ButtonDefaults.buttonColors(
-                containerColor = MaterialTheme.colorScheme.secondary,
-                contentColor = MaterialTheme.colorScheme.onPrimary,
-            ),
-            onClick = onNextPressed,
-        ) {
-            Text(
-                text = stringResource(R.string.onboarding_screen_4_cta),
-                fontSize = 20.sp,
-            )
-        }
-    }
-}
-
