+++ Android/EasyBudget/app/src/main/java/com/benoitletondor/easybudgetapp/view/onboarding/subviews/OnboardingPageCurrency.kt
@@ -1,0 +1,88 @@
+package com.benoitletondor.easybudgetapp.view.onboarding.subviews
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.colorResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.benoitletondor.easybudgetapp.R
+import com.benoitletondor.easybudgetapp.view.selectcurrency.SelectCurrencyView
+import kotlinx.coroutines.flow.StateFlow
+import java.util.Currency
+
+@Composable
+fun OnboardingPageCurrency(
+    contentPadding: PaddingValues,
+    userCurrencyFlow: StateFlow<Currency>,
+    onNextPressed: () -> Unit,
+) {
+    val selectedCurrency by userCurrencyFlow.collectAsState()
+
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(color = colorResource(R.color.secondary))
+            .padding(contentPadding),
+        horizontalAlignment = Alignment.CenterHorizontally,
+    ) {
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .weight(1f)
+                .padding(bottom = 20.dp),
+            horizontalAlignment = Alignment.CenterHorizontally,
+        ) {
+            Text(
+                modifier = Modifier.fillMaxWidth(),
+                text = stringResource(R.string.onboarding_screen_2_title),
+                color = Color.White,
+                fontSize = 30.sp,
+                textAlign = TextAlign.Center,
+                lineHeight = 36.sp,
+            )
+
+            Spacer(modifier = Modifier.height(30.dp))
+
+            Text(
+                modifier = Modifier.fillMaxWidth(),
+                text = stringResource(R.string.onboarding_screen_2_message),
+                color = Color.White,
+                fontSize = 20.sp,
+                textAlign = TextAlign.Center,
+            )
+
+            SelectCurrencyView(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .weight(weight = 1f, fill = false)
+                    .padding(horizontal = 20.dp, vertical = 30.dp)
+                    .background(color = Color.White),
+            )
+        }
+
+        Button(
+            onClick = onNextPressed,
+        ) {
+            Text(
+                text = stringResource(R.string.onboarding_screen_2_cta, selectedCurrency.symbol),
+                fontSize = 20.sp,
+            )
+        }
+    }
+}
