+++ Android/EasyBudget/app/src/main/java/com/benoitletondor/easybudgetapp/view/onboarding/subviews/OnboardingPagePushNotification.kt
@@ -1,0 +1,109 @@
+package com.benoitletondor.easybudgetapp.view.onboarding.subviews
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.colorResource
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.benoitletondor.easybudgetapp.R
+
+@Composable
+fun OnboardingPagePushNotifications(
+    contentPadding: PaddingValues,
+    onAcceptNotificationsPressed: () -> Unit,
+    onDenyNotificationsPressed: () -> Unit,
+) {
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(color = colorResource(R.color.secondary))
+            .padding(contentPadding),
+        horizontalAlignment = Alignment.CenterHorizontally,
+    ) {
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .weight(1f)
+                .verticalScroll(rememberScrollState())
+                .padding(bottom = 20.dp),
+            horizontalAlignment = Alignment.CenterHorizontally,
+        ) {
+            Image(
+                painter = painterResource(R.drawable.ic_baseline_notification_important_24),
+                contentDescription = null,
+            )
+
+            Spacer(modifier = Modifier.height(10.dp))
+
+            Text(
+                modifier = Modifier.fillMaxWidth(),
+                text = stringResource(R.string.onboarding_screen_push_permission_title),
+                color = Color.White,
+                fontSize = 30.sp,
+                textAlign = TextAlign.Center,
+                lineHeight = 36.sp,
+            )
+
+            Spacer(modifier = Modifier.height(20.dp))
+
+            Text(
+                modifier = Modifier.fillMaxWidth(),
+                text = stringResource(R.string.onboarding_screen_push_permission_message),
+                color = Color.White,
+                fontSize = 18.sp,
+                textAlign = TextAlign.Center,
+            )
+        }
+
+        Row(
+            modifier = Modifier.fillMaxWidth(),
+        ) {
+            Button(
+                modifier = Modifier.weight(0.5f),
+                onClick = onDenyNotificationsPressed,
+                colors = ButtonDefaults.buttonColors(
+                    containerColor = colorResource(R.color.accent_ripple),
+                    contentColor = colorResource(R.color.easy_budget_green_dark),
+                ),
+            ) {
+                Text(
+                    text = stringResource(R.string.onboarding_screen_push_permission_not_now_cta),
+                    fontSize = 20.sp,
+                )
+            }
+
+            Spacer(modifier = Modifier.width(10.dp))
+
+            Button(
+                modifier = Modifier.weight(0.5f),
+                onClick = onAcceptNotificationsPressed,
+            ) {
+                Text(
+                    text = stringResource(R.string.onboarding_screen_push_permission_accept_cta),
+                    fontSize = 20.sp,
+                )
+            }
+        }
+    }
+}
